first attemp at intermediate code generation
CHANGELOG:
- in parser.cup:
> fix for boolean expressions to 'act' like arithmetic expressions in
terms of operator precedence (NOT > AND > OR). Now the AST is generated
the same as the trees seen in class. Renamed some non-terminals and
created others.

- removed Arrays.asList from each array in SemanticAnalyzer and from
each line of code these where called (wasn't efficient), so I ended up
using an ArrayList for each operator list.

- created class IntermediateCode and Quadruple

- FIRST attempt at intermediate code generation:
> arithmetic expressions are now generated
> boolean expressions still in debugging, but expressions with OR in
them seem to work so far (generated code and propagates labels
corrrectly)
> working on simple IF statement code generation