merged initializations and declarations in .cup
CHANGELOG:

- now the grammar supports multiple declarations AND initializations in
one line. For example:

int x,y,z=0,k=1,i,j=z;

- the change above made semantic analysis easier, as well as
intermediate code. Besides, it should've always had it.

- now initializations have i-code representation

- fixed some bugs in label propagation (still one major bug to deal
with, see next change)

- there is a problem when propagating labels of blocks that are last in
their parent's children list. For example, if a 'main' has only an if
statement, (or anything that generates jumps), and such if statement has
children that generate jumps as well, the problem can be seen in the
last child of that if statement. If it generates jumps, those will be
generated without taking into account the 'jump when false' label of the
main's IF (they should both go to the same place, but they don't).